move selection functions to utils
